requirements.txt
Flask
flask-sqlalchemy
flask-login
flask-wtf
email-validator
psycopg2-binary
flask-migrate
python-dotenv
requests
stripe

